refactor: Extract Alert from Alerter
