+++ react/Alerter/index.jsx
@@ -1,10 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import Portal from '../Portal'
-import Button from '../Button'
-
-import styles from './styles.styl'
-import cx from 'classnames'
+import Alert from './Alert'
 
 const MINIMUM_ALERT_DURATION = 2000
 
@@ -26,111 +23,6 @@ const createStore = () => {
 }
 
 const store = createStore()
-
-export class Alert extends Component {
-  state = {
-    hidden: true
-  }
-
-  computeDuration() {
-    const words = this.props.message.split(/\W/).filter(Boolean)
-    return Math.max(MINIMUM_ALERT_DURATION, (words.length / 3) * 1000)
-  }
-
-  componentDidMount() {
-    this.closeTimer = setTimeout(
-      () => {
-        this.beginClosing()
-      },
-      this.props.duration === 'auto'
-        ? this.computeDuration()
-        : this.props.duration
-    )
-    // Delay to trigger CSS transition after the first render.
-    // Totally open for a better way to achieve this.
-    setTimeout(() => {
-      this.setState({ hidden: false })
-    }, 20)
-  }
-
-  beginClosing() {
-    this.base.addEventListener('transitionend', this.notifyClosed, false)
-    this.setState({ hidden: true })
-  }
-
-  notifyClosed = () => {
-    this.props.onClose()
-  }
-
-  componentWillUnmount() {
-    this.base.removeEventListener('transitionend', this.notifyClosed, false)
-    this.setState({ hidden: false })
-    if (this.closeTimer) {
-      clearTimeout(this.closeTimer)
-    }
-  }
-
-  close = () => {
-    if (this.closeTimer) clearTimeout(this.closeTimer)
-    this.beginClosing()
-  }
-
-  buttonAction = () => {
-    const { buttonAction } = this.props
-    // pass a way to dismiss the alerter from the button
-    if (typeof buttonAction === 'function') buttonAction(this.close)
-  }
-
-  render() {
-    const { message, type, buttonText } = this.props
-    const { hidden } = this.state
-    return (
-      <div
-        ref={c => (this.base = c)}
-        className={cx(
-          styles['c-alert'],
-          hidden ? styles['c-alert--hidden'] : ''
-        )}
-        role="alert"
-      >
-        <div
-          className={cx(styles['c-alert-wrapper'], styles[`c-alert--${type}`])}
-        >
-          <p>{message}</p>
-          {buttonText && (
-            <Button
-              onClick={this.buttonAction}
-              className={styles[`c-btn--alert-${type}`]}
-              label={buttonText}
-              size="tiny"
-            />
-          )}
-        </div>
-      </div>
-    )
-  }
-}
-
-Alert.propTypes = {
-  /** @type string - Controls the style of the error */
-  type: PropTypes.oneOf(['success', 'info', 'error']),
-  /** @type {string} - Message to display */
-  message: PropTypes.string.isRequired,
-  /** @type {function} - Callback when is dismissed */
-  onClose: PropTypes.func,
-  /** @type {function} - Text of the button, if absent, no button is displayed */
-  buttonText: PropTypes.string,
-  /** @type {function} - Callback when clicking on the button */
-  buttonAction: PropTypes.func
-}
-
-Alert.defaultProps = {
-  type: 'info',
-  onClose: () => {},
-  buttonText: undefined,
-  buttonAction: () => {},
-  duration: 'auto'
-}
 
 class Alerter extends Component {
   state = {
@@ -202,6 +94,7 @@ class Alerter extends Component {
             buttonText={notif.options && notif.options.buttonText}
             buttonAction={notif.options && notif.options.buttonAction}
             duration={notif.options && notif.options.duration}
+            minDuration={MINIMUM_ALERT_DURATION}
           />
         ))}
       </Portal>
